Move encodeActCall to lib
